integrate local executor
