+++ tests/agent_server/test_server_agent.py
@@ -967,143 +967,6 @@ async def test_fail(resource_container, client, agent, environment, clienthelper
     assert states["test::Resource[agent1,key=key5],v=%d" % version] == "skipped"
 
 
-async def test_wait(resource_container, client, clienthelper, environment, server, no_agent_backoff, async_finalizer):
-    """
-    If this test fail due to timeout,
-    this is probably due to the mechanism in the agent that prevents pulling resources in very rapid succession.
-
-    If the test server is slow, a get_resources call takes a long time,
-    this makes the back-off longer
-
-    this test deploys two models in rapid successions, if the server is slow, this may fail due to the back-off
-    """
-    resource_container.Provider.reset()
-
-    env_id = environment
-
-    # setup agent
-    agent = Agent(hostname="node1", environment=env_id, agent_map={"agent1": "localhost"}, code_loader=False, poolsize=10)
-    await agent.add_end_point_name("agent1")
-    await agent.start()
-    async_finalizer(agent.stop)
-
-    # wait for agent
-    await retry_limited(lambda: len(server.get_slice(SLICE_SESSION_MANAGER)._sessions) == 1, 10)
-
-    # set the deploy environment
-    resource_container.Provider.set("agent1", "key", "value")
-
-    async def make_version():
-        version = await clienthelper.get_version()
-
-        resources = [
-            {
-                "key": "key",
-                "value": "value",
-                "id": "test::Wait[agent1,key=key],v=%d" % version,
-                "requires": [],
-                "purged": False,
-                "send_event": False,
-            },
-            {
-                "key": "key2",
-                "value": "value",
-                "id": "test::Resource[agent1,key=key2],v=%d" % version,
-                "requires": ["test::Wait[agent1,key=key],v=%d" % version],
-                "purged": False,
-                "send_event": False,
-            },
-            {
-                "key": "key3",
-                "value": "value",
-                "id": "test::Resource[agent1,key=key3],v=%d" % version,
-                "requires": [],
-                "purged": False,
-                "send_event": False,
-            },
-            {
-                "key": "key4",
-                "value": "value",
-                "id": "test::Resource[agent1,key=key4],v=%d" % version,
-                "requires": ["test::Resource[agent1,key=key3],v=%d" % version],
-                "purged": False,
-                "send_event": False,
-            },
-            {
-                "key": "key5",
-                "value": "value",
-                "id": "test::Resource[agent1,key=key5],v=%d" % version,
-                "requires": ["test::Resource[agent1,key=key4],v=%d" % version, "test::Wait[agent1,key=key],v=%d" % version],
-                "purged": False,
-                "send_event": False,
-            },
-        ]
-        return version, resources
-
-    logger.info("setup done")
-
-    version1, resources = await make_version()
-    result = await client.put_version(
-        tid=env_id, version=version1, resources=resources, unknowns=[], version_info={}, compiler_version=get_compiler_version()
-    )
-    assert result.code == 200
-
-    logger.info("first version pushed")
-
-    # deploy and wait until one is ready
-    result = await client.release_version(env_id, version1, True, const.AgentTriggerMethod.push_full_deploy)
-    assert result.code == 200
-
-    logger.info("first version released")
-
-    await wait_for_n_deployed_resources(client, env_id, version1, n=2)
-
-    generation_version1 = agent._instances["agent1"]._nq.generation
-
-    result = await client.get_version(environment, version1)
-    assert result.code == 200
-    assert result.result["model"]["done"] == 2
-
-    logger.info("first version, 2 resources deployed")
-
-    version2, resources = await make_version()
-    result = await client.put_version(
-        tid=env_id, version=version2, resources=resources, unknowns=[], version_info={}, compiler_version=get_compiler_version()
-    )
-    assert result.code == 200
-
-    logger.info("second version pushed %f", time.time())
-
-    # deploy and wait until done
-    result = await client.release_version(env_id, version2, True, const.AgentTriggerMethod.push_full_deploy)
-    assert result.code == 200
-
-    logger.info("second version released")
-
-    # Wait until the agent is deploying version2. Otherwise the call to `wait_for_done_with_waiters()`
-    # might notify a test::Wait resource of version1, instead of the resources of version2. This could
-    # cause problem when the resource states of version1 are checked later on in this test case.
-    await retry_limited(lambda: agent._instances["agent1"]._nq.generation != generation_version1, timeout=10)
-    await resource_container.wait_for_done_with_waiters(client, env_id, version2)
-
-    logger.info("second version complete")
-
-    result = await client.get_version(env_id, version2)
-    assert result.code == 200
-    for x in result.result["resources"]:
-        assert x["status"] == const.ResourceState.deployed.name
-
-    result = await client.get_version(env_id, version1)
-    assert result.code == 200
-    states = {x["id"]: x["status"] for x in result.result["resources"]}
-
-    assert states["test::Wait[agent1,key=key],v=%d" % version1] == const.ResourceState.deployed.name
-    assert states["test::Resource[agent1,key=key2],v=%d" % version1] == const.ResourceState.available.name
-    assert states["test::Resource[agent1,key=key3],v=%d" % version1] == const.ResourceState.deployed.name
-    assert states["test::Resource[agent1,key=key4],v=%d" % version1] == const.ResourceState.deployed.name
-    assert states["test::Resource[agent1,key=key5],v=%d" % version1] == const.ResourceState.available.name
-
-
 async def test_multi_instance(resource_container, client, clienthelper, server, environment, no_agent_backoff, async_finalizer):
     """
     Test for multi threaded deploy
@@ -1118,7 +981,6 @@ async def test_multi_instance(resource_container, client, clienthelper, server, 
         environment=env_id,
         agent_map={"agent1": "localhost", "agent2": "localhost", "agent3": "localhost"},
         code_loader=False,
-        poolsize=1,
     )
     await agent.add_end_point_name("agent1")
     await agent.add_end_point_name("agent2")
@@ -2566,7 +2428,8 @@ async def test_s_full_deploy_waits_for_incremental_deploy(
         await asyncio.sleep(0.1)
 
     # cache has 1 version in flight
-    assert len(myagent_instance._cache.counterforVersion) == 1
+    executor_instance = agent.executor_manager.executors["agent1"]
+    assert len(executor_instance._cache.counterforVersion) == 1
 
     version2 = await clienthelper.get_version()
     resources_version_2 = get_resources(version2, "value3")
@@ -2579,7 +2442,7 @@ async def test_s_full_deploy_waits_for_incremental_deploy(
 
     # cache has no versions in flight
     # for issue #1883
-    assert not myagent_instance._cache.counterforVersion
+    assert not executor_instance._cache.counterforVersion
 
     # Incremental deploy
     #   * All resources are deployed successfully:
@@ -2811,7 +2674,8 @@ async def test_s_periodic_Vs_full(
         await asyncio.sleep(0.1)
 
     # cache has 1 version in flight
-    assert len(myagent_instance._cache.counterforVersion) == 1
+    executor_instance = agent.executor_manager.executors["agent1"]
+    assert len(executor_instance._cache.counterforVersion) == 1
 
     version2 = await clienthelper.get_version()
     resources_version_2 = get_resources(version2, "value3")
@@ -2824,7 +2688,8 @@ async def test_s_periodic_Vs_full(
     await resource_container.wait_for_done_with_waiters(client, environment, versions[action.wait_for])
     # cache has no versions in flight
     # for issue #1883
-    assert not myagent_instance._cache.counterforVersion
+
+    assert not executor_instance._cache.counterforVersion
 
     log_contains(caplog, "inmanta.agent.agent.agent1", logging.INFO, action.msg)
